pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "peertranslate"
version = "0.2.0"
description = "Verified research paper translation for every language"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Muhammed Muminul Hoque"}
]
keywords = [
    "research-paper-translation",
    "academic-translation",
    "bengali",
    "multilingual",
    "gemini-api",
    "back-translation-verification",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.1",
    "sse-starlette>=2.1.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "PyMuPDF>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[project.urls]
Homepage = "https://github.com/muminul-hoque/PeerTranslate"
Repository = "https://github.com/muminul-hoque/PeerTranslate"
Issues = "https://github.com/muminul-hoque/PeerTranslate/issues"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
